Add files via upload
There are millions of reviews on online shopping app for products to services. The generation of reviews by users is an important part of how opinions are formed when somebody want to make a decision to shop, it is important to classified them in helpfulness and unhelpfulness. This would help users to have a clear idea about the products which will eventually give better buying decisions. The purpose of this project is to get the most helpful product reviews based on review in form of texts and ratings in the dataset.